requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.4
Django==5.2.11
django-chartjs==2.3.0
django-jazzmin==3.0.3
Faker==40.5.1
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
reportlab==4.4.10
sqlparse==0.5.5
typing_extensions==4.15.0
